refactor(Summary): quickfix update to Summary to display AddAnother questions and answers in base format
