+++ src/Factories/Schema/SchemaFactory.cs
@@ -2,10 +2,8 @@ using System.Threading.Tasks;
 using form_builder.Configuration;
 using form_builder.Enum;
 using form_builder.Extensions;
-using form_builder.Factories.Transform.AddAnother;
 using form_builder.Factories.Transform.Lookups;
 using form_builder.Factories.Transform.ReusableElements;
-using form_builder.Helpers.PageHelpers;
 using form_builder.Models;
 using form_builder.Providers.SchemaProvider;
 using form_builder.Providers.StorageProvider;
