+++ src/Helpers/PageHelpers/IPageHelper.cs
@@ -9,6 +9,8 @@ namespace form_builder.Helpers.PageHelpers
     {
         Task<FormBuilderViewModel> GenerateHtml(Page page, Dictionary<string, dynamic> viewModel, FormSchema baseForm, string guid, FormAnswers formAnswers, List<object> results = null);
 
+        (FormSchema dynamicFormSchema, Page dynamicCurrentPage) GetDynamicFormSchema(Page currentPage, string guid);
+
         void RemoveAnswers(Dictionary<string, dynamic> answersToRemove, string guid, string path);
 
         void SaveAnswers(Dictionary<string, dynamic> viewModel, string guid, string form, IEnumerable<CustomFormFile> files, bool isPageValid, bool appendMultipleFileUploadParts = false);
